+++ Lilu/Sources/kern_api.cpp
@@ -8,6 +8,7 @@
 #include <Headers/kern_config.hpp>
 #include <PrivateHeaders/kern_config.hpp>
 #include <Headers/kern_api.hpp>
+#include <Headers/kern_devinfo.hpp>
 
 #include <IOKit/IOLib.h>
 #include <IOKit/IORegistryEntry.h>
@@ -235,6 +236,8 @@ void LiluAPI::processPatcherLoadCallbacks(KernelPatcher &patcher) {
 	IOLockLock(access);
 	apiRequestsOver = true;
 	IOLockUnlock(access);
+
+	DeviceInfo::create_cached();
 
 	// Process the callbacks
 	for (size_t i = 0; i < patcherLoadedCallbacks.size(); i++) {