+++ Changelog.md
@@ -6,6 +6,8 @@ Lilu Changelog
 - Added `external-audio` property to ignore PCI audio cards
 - Added in-memory symbol solving for 11.0
 - Fixed accidentally solving stabs instead of normal symbols
+- Added device publishing API to monitor device startup
+- Added DeviceInfo caching for improved performance
 
 #### v1.4.5
 - Fixed newer CPU generation detection